% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{sBG.log.likelihood}
\alias{sBG.log.likelihood}
\title{Shifted Beta Geometric Model}
\usage{
sBG.log.likelihood(params)
}
\arguments{
\item{alpha}{Alpha (starts as 1)}

\item{beta}{Beta (starts as 1)}
}
\description{
This function allows you to implement as shifted beta geometric model
}
\examples{
sBG.log.likelihood()

}
\keyword{beta}
\keyword{geometric}
\keyword{shifted}
