% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{sBG.table}
\alias{sBG.table}
\title{Helper function 1}
\usage{
sBG.table(params, data)
}
\arguments{
\item{alpha}{Alpha (starts as 1)}

\item{beta}{Beta (starts as 1)}
}
\description{
This function allows you to implement as shifted beta geometric model
}
\examples{
sBG.table()

}
\keyword{beta}
\keyword{geometric}
\keyword{shifted}
