% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/churn.R
\name{sbg_function}
\alias{sbg_function}
\title{Helper function 2}
\usage{
sbg_function(params)
}
\arguments{
\item{alpha}{Alpha (starts as 1)}

\item{beta}{Beta (starts as 1)}
}
\description{
This function allows you to implement as shifted beta geometric model
}
\examples{
sbg_function()

}
\keyword{beta}
\keyword{geometric}
\keyword{shifted}
